dreambooks_posts_generator/requirements.txt
asgiref==3.6.0
Django==4.2
numpy==1.24.2
Pillow==9.5.0
reportlab==3.6.12
sqlparse==0.4.3
tzdata==2023.3
